# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sales": {
      "post": {
        "operationId": "SalesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "sale"
        ]
      },
      "get": {
        "operationId": "SalesController_findAll",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "customerName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "start",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SaleRespondeDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "sale"
        ]
      }
    },
    "/sales/{id}": {
      "get": {
        "operationId": "SalesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "sale"
        ]
      },
      "patch": {
        "operationId": "SalesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSaleDto"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "sale"
        ]
      },
      "delete": {
        "operationId": "SalesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaleRespondeDto"
                }
              }
            }
          }
        },
        "tags": [
          "sale"
        ]
      }
    },
    "/sales/resport/{id}": {
      "get": {
        "operationId": "SalesController_generatePdf",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "sale"
        ]
      }
    },
    "/sales/resports": {
      "get": {
        "operationId": "SalesController_generateReport",
        "parameters": [
          {
            "name": "id",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "customerName",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "stage",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": false,
            "in": "query",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "customerId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "start",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "end",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "sale"
        ]
      }
    },
...
  },
  "info": {
    "title": "MI PROFE",
    "description": "The MI PROFE-API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "api"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    },
    "schemas": {
...
      "IdCreateEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          }
        }
      },
...
      "CreateSaleProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "amount": {
            "type": "string"
          },
          "price": {
            "type": "string"
          },
          "subtotal": {
            "type": "string"
          },
          "product": {
            "$ref": "#/components/schemas/IdCreateEntity"
          }
        },
        "required": [
          "amount",
          "price",
          "subtotal",
          "product"
        ]
      },
      "CreateSaleDto": {
        "type": "object",
        "properties": {
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "total": {
            "type": "number"
          },
          "customer": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "saleProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateSaleProductDto"
            }
          }
        },
        "required": [
          "date",
          "stage",
          "total",
          "customer",
          "saleProducts"
        ]
      },
      "Customer": {
        "type": "object",
        "properties": {}
      },
      "SaleRespondeDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "total": {
            "type": "number"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "saleProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateSaleProductDto"
            }
          }
        },
        "required": [
          "id",
          "date",
          "stage",
          "total",
          "customer",
          "saleProducts"
        ]
      },
      "ReportsResponseDto": {
        "type": "object",
        "properties": {
          "reportUrl": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "reportUrl",
          "name"
        ]
      },
      "UpdateSaleDto": {
        "type": "object",
        "properties": {
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "note": {
            "type": "string"
          },
          "stage": {
            "type": "string"
          },
          "total": {
            "type": "number"
          },
          "customer": {
            "$ref": "#/components/schemas/IdCreateEntity"
          },
          "saleProducts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateSaleProductDto"
            }
          }
        }
      },
...
    }
  }
}